handle select all logic
